requirements.txt
requests==2.32.3
Flask==2.0.2
python-dotenv==0.19.2